Handle IGNORE trap handlers
