--- .travis.yml
@@ -1,11 +1,0 @@
-rvm:
-  - 1.9.2
-  - 1.9.3
-  - 2.0.0
-  - 2.1
-  - 2.2
-  - ruby-head
-  - rbx-2
-  - jruby-9.0.0.0
-
-sudo: false
