+++ Gemfile
@@ -1,4 +1,5 @@
+# frozen_string_literal: true
+
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in multitrap.gemspec
 gemspec
