+++ Rakefile
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'rspec/core/rake_task'
 RSpec::Core::RakeTask.new(:spec)
